rename NotificationLine.* files to gui_NotificationLine.*
