+++ src/code/launcher/gui_launcher.h
@@ -3,7 +3,7 @@
 //
 #pragma once
 
-#include "NotificationLine.h"
+#include "gui_NotificationLine.h"
 #include "../gui/gui_screen.h"
 #include "ps_obj.h"
 #include "ps_settings_back.h"
